Update GemCutting.lua
Update to gem cutting tab to replace the wrath stuff. The links do work on beta but using this on live will just show dead links as the new gems are not in yet.

As I'm not very confident in Lua yet, I have not changed the code to add in new parts to the gui but only changed all the item numbers and rearranged the layout for the MOP item numbers.

I'm just doing this to encourage Tekkub to hopefully update the addon as I've looked up the item numbers for him. If there's interest then i'm thinking about jigging up the inscription and milling sections as well.