Update GemCutting.lua
Updated gemcutting.lua to put back the wrath stuff I overwrote and resort the order of the tabs. Also added in a couple of items I missed originally. I haven't been able to get the item to make one serpent gem out of 10 shards going yet but i'm working on it.